# DebtManagement/src/main/java/vn/diepgia/mchis/DebtManagement/responses/AuthenticationResponse.java
package vn.diepgia.mchis.DebtManagement.responses;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;

@Data
@AllArgsConstructor
@Builder
public class AuthenticationResponse {
    private String token;
}